+++ front/lib/actions/process.ts
@@ -4,12 +4,14 @@ import {
   DEFAULT_PROCESS_ACTION_NAME,
   DEFAULT_SEARCH_LABELS_ACTION_NAME,
 } from "@app/lib/actions/constants";
+import type {
+  DataSourceConfiguration,
+  RetrievalTimeframe,
+} from "@app/lib/actions/retrieval";
 import {
   applyDataSourceFilters,
-  parseTimeFrame,
   retrievalAutoTimeFrameInputSpecification,
   retrievalTagsInputSpecification,
-  timeFrameFromNow,
 } from "@app/lib/actions/retrieval";
 import { runActionStreamed } from "@app/lib/actions/server";
 import type { ExtractActionBlob } from "@app/lib/actions/types";
@@ -20,10 +22,6 @@ import type {
 } from "@app/lib/actions/types/agent";
 import type { BaseActionRunParams } from "@app/lib/actions/types/base";
 import { BaseActionConfigurationServerRunner } from "@app/lib/actions/types/base";
-import type {
-  DataSourceConfiguration,
-  RetrievalTimeframe,
-} from "@app/lib/actions/types/retrieval";
 import { constructPromptMultiActions } from "@app/lib/api/assistant/generation";
 import { getSupportedModelConfig } from "@app/lib/assistant";
 import type { Authenticator } from "@app/lib/auth";
@@ -40,7 +38,7 @@ import type {
   TimeFrame,
   UserMessageType,
 } from "@app/types";
-import { Ok } from "@app/types";
+import { Ok, parseTimeFrame, timeFrameFromNow } from "@app/types";
 
 export const PROCESS_SCHEMA_ALLOWED_TYPES = [
   "string",
