+++ front/lib/actions/retrieval.ts
@@ -6,6 +6,7 @@ import {
   DEFAULT_SEARCH_LABELS_ACTION_NAME,
 } from "@app/lib/actions/constants";
 import { runActionStreamed } from "@app/lib/actions/server";
+import type { ExtractActionBlob } from "@app/lib/actions/types";
 import { BaseAction } from "@app/lib/actions/types";
 import type {
   ActionConfigurationType,
@@ -13,18 +14,6 @@ import type {
 } from "@app/lib/actions/types/agent";
 import type { BaseActionRunParams } from "@app/lib/actions/types/base";
 import { BaseActionConfigurationServerRunner } from "@app/lib/actions/types/base";
-import type {
-  DataSourceConfiguration,
-  RetrievalDocumentType,
-} from "@app/lib/actions/types/retrieval";
-import type {
-  RetrievalActionType,
-  RetrievalConfigurationType,
-  RetrievalDocumentChunkType,
-  RetrievalErrorEvent,
-  RetrievalParamsEvent,
-  RetrievalSuccessEvent,
-} from "@app/lib/actions/types/retrieval";
 import { actionRefsOffset, getRetrievalTopK } from "@app/lib/actions/utils";
 import { getRefs } from "@app/lib/api/assistant/citations";
 import type { Authenticator } from "@app/lib/auth";
@@ -37,101 +26,151 @@ import { RetrievalDocumentResource } from "@app/lib/resources/retrieval_document
 import { generateRandomModelSId } from "@app/lib/resources/string_ids";
 import logger from "@app/logger/logger";
 import type {
+  ConnectorProvider,
+  DataSourceViewType,
   FunctionCallType,
   FunctionMessageTypeModel,
   ModelId,
   Result,
+  TagsFilter,
   TimeFrame,
 } from "@app/types";
-import { Ok } from "@app/types";
+import { Ok, parseTimeFrame, timeFrameFromNow } from "@app/types";
+
+export type DataSourceFilter = {
+  parents: { in: string[]; not: string[] } | null;
+  tags?: TagsFilter;
+};
+
+export type DataSourceConfiguration = {
+  workspaceId: string;
+  dataSourceViewId: string;
+  filter: DataSourceFilter;
+};
+
 /**
- * TimeFrame parsing
+ * Retrieval configuration
  */
 
-// Attempts to parse a string representation of the time frame of the form `{k}{unit}` or `all`
-// where {k} is a number and {unit} is one of `d`, `w`, `m`, `y` for day, week, month, year.
-export function parseTimeFrame(raw: string): TimeFrame | null {
-  const r = raw.trim().toLowerCase();
-  if (r === "all") {
-    return null;
+// Retrieval specifies a list of data sources (with possible parent filtering, possible "all" data
+// sources), a query ("auto" generated by the model "none", no query, `TemplatedQuery`, fixed
+// query), a relative time frame ("auto" generated by the model, "none" no time filtering
+// `TimeFrame`) which applies to all data sources, and a top_k parameter.
+//
+// `query` and `relativeTimeFrame` will be used to generate the inputs specification for the model
+// in charge of generating the action inputs. The results will be used along with `topK` and
+// `dataSources` to query the data.
+export type RetrievalQuery = "auto" | "none";
+export type RetrievalTimeframe = "auto" | "none" | TimeFrame;
+export type RetrievalConfigurationType = {
+  id: ModelId;
+  sId: string;
+
+  type: "retrieval_configuration";
+
+  dataSources: DataSourceConfiguration[];
+  query: RetrievalQuery;
+  relativeTimeFrame: RetrievalTimeframe;
+  topK: number | "auto";
+
+  name: string;
+  description: string | null;
+};
+
+/**
+ * Retrieval action
+ */
+
+export interface RetrievalDocumentChunkType {
+  offset: number;
+  score: number | null;
+  text: string;
+}
+
+export interface RetrievalDocumentType {
+  chunks: RetrievalDocumentChunkType[];
+  documentId: string;
+  dataSourceView: DataSourceViewType | null;
+  id: ModelId;
+  reference: string; // Short random string so that the model can refer to the document.
+  score: number | null;
+  sourceUrl: string | null;
+  tags: string[];
+  timestamp: number;
+}
+
+type ConnectorProviderDocumentType =
+  | Exclude<ConnectorProvider, "webcrawler">
+  | "document";
+
+export function getProviderFromRetrievedDocument(
+  document: RetrievalDocumentType
+): ConnectorProviderDocumentType {
+  if (document.dataSourceView) {
+    if (document.dataSourceView.dataSource.connectorProvider === "webcrawler") {
+      return "document";
+    }
+    return document.dataSourceView.dataSource.connectorProvider || "document";
+  }
+  return "document";
+}
+
+export function getTitleFromRetrievedDocument(
+  document: RetrievalDocumentType
+): string {
+  const provider = getProviderFromRetrievedDocument(document);
+
+  if (provider === "slack") {
+    for (const t of document.tags) {
+      if (t.startsWith("channelName:")) {
+        return `#${t.substring(12)}`;
+      }
+    }
   }
 
-  const m = r.match(/^(\d+)([hdwmy])$/);
-  if (!m) {
-    return null;
+  for (const t of document.tags) {
+    if (t.startsWith("title:")) {
+      return t.substring(6);
+    }
   }
 
-  const duration = parseInt(m[1], 10);
-  if (isNaN(duration)) {
-    return null;
-  }
+  return document.documentId;
+}
 
-  let unit: TimeFrame["unit"];
-  switch (m[2]) {
-    case "h":
-      unit = "hour";
-      break;
-    case "d":
-      unit = "day";
-      break;
-    case "w":
-      unit = "week";
-      break;
-    case "m":
-      unit = "month";
-      break;
-    case "y":
-      unit = "year";
-      break;
-    default:
-      return null;
-  }
+// Event sent during retrieval with the finalized query used to retrieve documents.
+type RetrievalParamsEvent = {
+  type: "retrieval_params";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  dataSources: DataSourceConfiguration[];
+  action: RetrievalActionType;
+};
 
-  return {
-    duration,
-    unit,
+type RetrievalErrorEvent = {
+  type: "retrieval_error";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  error: {
+    code: string;
+    message: string;
   };
-}
+};
 
-// Turns a TimeFrame into a number of milliseconds from now.
-export function timeFrameFromNow(timeFrame: TimeFrame): number {
-  const now = Date.now();
+type RetrievalSuccessEvent = {
+  type: "retrieval_success";
+  created: number;
+  configurationId: string;
+  messageId: string;
+  action: RetrievalActionType;
+};
 
-  switch (timeFrame.unit) {
-    case "hour":
-      return now - timeFrame.duration * 60 * 60 * 1000;
-    case "day":
-      return now - timeFrame.duration * 24 * 60 * 60 * 1000;
-    case "week":
-      return now - timeFrame.duration * 7 * 24 * 60 * 60 * 1000;
-    case "month":
-      return now - timeFrame.duration * 30 * 24 * 60 * 60 * 1000;
-    case "year":
-      return now - timeFrame.duration * 365 * 24 * 60 * 60 * 1000;
-    default:
-      ((x: never) => {
-        throw new Error(`Unexpected time frame unit ${x}`);
-      })(timeFrame.unit);
-  }
-}
+export type RetrievalActionRunningEvents = RetrievalParamsEvent;
 
-interface RetrievalActionBlob {
-  id: ModelId; // AgentRetrievalAction.
-  agentMessageId: ModelId;
-  params: {
-    relativeTimeFrame: TimeFrame | null;
-    query: string | null;
-    topK: number;
-    tagsIn: string[] | null;
-    tagsNot: string[] | null;
-  };
-  functionCallId: string | null;
-  functionCallName: string | null;
-  documents: RetrievalDocumentType[] | null;
-  step: number;
-}
+type RetrievalActionBlob = ExtractActionBlob<RetrievalActionType>;
 
-export class RetrievalAction extends BaseAction {
+export class RetrievalActionType extends BaseAction {
   readonly agentMessageId: ModelId;
   readonly params: {
     relativeTimeFrame: TimeFrame | null;
@@ -147,7 +186,7 @@ export class RetrievalAction extends BaseAction {
   readonly type = "retrieval_action";
 
   constructor(blob: RetrievalActionBlob) {
-    super(blob.id, "retrieval_action");
+    super(blob.id, blob.type);
 
     this.agentMessageId = blob.agentMessageId;
     this.params = blob.params;
@@ -491,7 +530,7 @@ export class RetrievalConfigurationServerRunner extends BaseActionConfigurationS
       configurationId: agentConfiguration.sId,
       messageId: agentMessage.sId,
       dataSources: actionConfiguration.dataSources,
-      action: new RetrievalAction({
+      action: new RetrievalActionType({
         id: action.id,
         agentMessageId: agentMessage.agentMessageId,
         params: {
@@ -505,6 +544,8 @@ export class RetrievalConfigurationServerRunner extends BaseActionConfigurationS
         functionCallName: action.functionCallName,
         documents: null,
         step: action.step,
+        type: "retrieval_action",
+        generatedFiles: [],
       }),
     };
 
@@ -750,7 +791,7 @@ export class RetrievalConfigurationServerRunner extends BaseActionConfigurationS
       created: Date.now(),
       configurationId: agentConfiguration.sId,
       messageId: agentMessage.sId,
-      action: new RetrievalAction({
+      action: new RetrievalActionType({
         id: action.id,
         agentMessageId: agentMessage.agentMessageId,
         params: {
@@ -764,6 +805,8 @@ export class RetrievalConfigurationServerRunner extends BaseActionConfigurationS
         functionCallName: action.functionCallName,
         documents: documents.map((d) => d.toJSON(auth)),
         step: action.step,
+        type: "retrieval_action",
+        generatedFiles: [],
       }),
     };
   }
@@ -927,7 +970,7 @@ export async function retrievalActionTypesFromAgentMessageIds(
     });
 
     actions.push(
-      new RetrievalAction({
+      new RetrievalActionType({
         id: action.id,
         agentMessageId: action.agentMessageId,
         params: {
@@ -941,6 +984,8 @@ export async function retrievalActionTypesFromAgentMessageIds(
         functionCallName: action.functionCallName,
         documents,
         step: action.step,
+        type: "retrieval_action",
+        generatedFiles: [],
       })
     );
   }
