+++ front/lib/actions/runners.ts
@@ -7,6 +7,7 @@ import type { ProcessConfigurationType } from "@app/lib/actions/process";
 import { ProcessConfigurationServerRunner } from "@app/lib/actions/process";
 import type { ReasoningConfigurationType } from "@app/lib/actions/reasoning";
 import { ReasoningConfigurationServerRunner } from "@app/lib/actions/reasoning";
+import type { RetrievalConfigurationType } from "@app/lib/actions/retrieval";
 import { RetrievalConfigurationServerRunner } from "@app/lib/actions/retrieval";
 import { SearchLabelsConfigurationServerRunner } from "@app/lib/actions/search_labels";
 import type { TablesQueryConfigurationType } from "@app/lib/actions/tables_query";
@@ -18,7 +19,6 @@ import type {
   BaseActionConfigurationStaticMethods,
 } from "@app/lib/actions/types/base";
 import type { ConversationIncludeFileConfigurationType } from "@app/lib/actions/types/conversation/include_file";
-import type { RetrievalConfigurationType } from "@app/lib/actions/types/retrieval";
 import type { SearchLabelsConfigurationType } from "@app/lib/actions/types/search_labels";
 import type { WebsearchConfigurationType } from "@app/lib/actions/types/websearch";
 import { WebsearchConfigurationServerRunner } from "@app/lib/actions/websearch";
